+++ dinero/_dinero.py
@@ -17,12 +17,15 @@ The module contains the following methods:
 """
 
 import json
-from decimal import Decimal, InvalidOperation, getcontext
+from decimal import Decimal, getcontext
 from typing import Any
 
 from ._utils import DecimalEncoder
 from .exceptions import DifferentCurrencyError, InvalidOperationError
 from .types import Currency, OperationType
+from ._validators import Validate
+
+validator = Validate()
 
 
 class Base:
@@ -32,7 +35,7 @@ class Base:
         self.amount = amount
         self.currency = currency
 
-        _validate_dinero_amount(amount)
+        validator.validate_dinero_amount(amount)
 
     @property
     def symbol(self):
@@ -116,7 +119,7 @@ class Operations(Utils):
     """All the operations supported between Dinero objects."""
 
     def __add__(self, addend: "OperationType | Dinero") -> "Dinero":
-        _validate_addition_subtraction_amount(addend)
+        validator.validate_addition_subtraction_amount(addend)
         addend_obj = self._get_instance(addend)
         total = self._normalize() + addend_obj._normalize()
         return Dinero(str(total), self.currency)
@@ -125,19 +128,19 @@ class Operations(Utils):
         return self
 
     def __sub__(self, subtrahend: "OperationType | Dinero") -> "Dinero":
-        _validate_addition_subtraction_amount(subtrahend)
+        validator.validate_addition_subtraction_amount(subtrahend)
         subtrahend_obj = self._get_instance(subtrahend)
         total = self._normalize() - subtrahend_obj._normalize()
         return Dinero(str(total), self.currency)
 
     def __mul__(self, multiplicand: int | float | Decimal) -> "Dinero":
-        _validate_multiplication_division_amount(multiplicand)
+        validator.validate_multiplication_division_amount(multiplicand)
         multiplicand_obj = self._get_instance(multiplicand)
         total = self._normalize() * multiplicand_obj._normalize()
         return Dinero(str(total), self.currency)
 
     def __truediv__(self, divisor: int | float | Decimal) -> "Dinero":
-        _validate_multiplication_division_amount(divisor)
+        validator.validate_multiplication_division_amount(divisor)
         divisor_obj = self._get_instance(divisor)
         total = self._normalize() / divisor_obj._normalize()
         return Dinero(str(total), self.currency)
@@ -556,56 +559,3 @@ class Dinero(Operations, Base):
     def __str__(self):
         formatted_output = self.format()
         return f"{formatted_output}"
-
-
-def _validate_dinero_amount(amount: int | float | str | Decimal) -> None:
-    """Validate that the amount passed to Dinero is of valid type
-
-    Args:
-        amount (str, int, float, Decimal, Dinero)
-
-    Raises:
-        InvalidOperationError: An operation between unsupported types was executed.
-    """
-    try:
-        if not isinstance(amount, (int, float, str, Decimal, Dinero)):
-            raise InvalidOperationError(InvalidOperationError.operation_msg)
-
-        Decimal(amount)
-
-    except (ValueError, InvalidOperation):
-        raise InvalidOperationError(InvalidOperationError.operation_msg)
-
-
-def _validate_addition_subtraction_amount(amount: "OperationType | Dinero") -> None:
-    """Validate that the amount passed to an addition or subtraction is of valid type.
-
-    Args:
-        amount (str, int, float, Decimal, Dinero)
-
-    Raises:
-        InvalidOperationError: An operation between unsupported types was executed.
-    """
-    try:
-        if not isinstance(amount, (int, float, str, Decimal, Dinero)):
-            raise InvalidOperationError(InvalidOperationError.operation_msg)
-
-    except (ValueError, InvalidOperation):
-        raise InvalidOperationError(InvalidOperationError.operation_msg)
-
-
-def _validate_multiplication_division_amount(number: int | float | Decimal) -> None:
-    """Validate that the number passed to a multiplication and division is of valid type.
-
-    Args:
-        amount (int, float, Decimal)
-
-    Raises:
-        InvalidOperationError: An operation between unsupported types was executed.
-    """
-    try:
-        if not isinstance(number, (int, float, Decimal)):
-            raise InvalidOperationError(InvalidOperationError.operation_msg)
-
-    except (ValueError, InvalidOperation):
-        raise InvalidOperationError(InvalidOperationError.operation_msg)