+++ CHANGELOG.md
@@ -2,11 +2,12 @@
 
 ## [Unreleased](https://github.com/wilfredinni/dinero/compare/0.1.5...master)
 
-- Fix types for multiply and divide methods.
+- Fix Type Hints for multiply and divide methods.
+- Move validators to their own module.
 
 ## [0.1.5](https://github.com/wilfredinni/dinero/releases/tag/0.1.5) (2022-11-03)
 
--  It is no longer possible to compare a `Dinero` instance to objects of other types ([796b9e2](https://github.com/wilfredinni/dinero/commit/796b9e2e1f344f20be14edffc9a0579192d9a93b)).
+- It is no longer possible to compare a `Dinero` instance to objects of other types ([796b9e2](https://github.com/wilfredinni/dinero/commit/796b9e2e1f344f20be14edffc9a0579192d9a93b)).
 - Multiplication and division can be performed only with `int`, `float` and `Decimal` types ([326ad3a](https://github.com/wilfredinni/dinero/commit/326ad3a007c625e22fba2b265c093ee13f8a90d2)).
 - Added GitHub community files: SECURITY, CHANGELOG and CONTRIBUTING.
 - Added coverage and code quality checks.